tipc: Corrections to configuration code
commit f2f9e9b13f900469fba68deca1caa036d7a7824d from
git://tipc.cslab.ericsson.net/pub/git/people/allan/tipc.git

Modifies TIPC's configuration-related code to standardize the types
that signify the size of items used in configuration messages;
all size values (whether 16 bit, 32 bit, or size_t) are now unsigned.
These changes correct several places where a size value could be
accidentally truncated and also eliminate warnings where a mix of
signed and unsigned values are involved.

Integrated-by: Howard Xu <hao.xu@windriver.com>

+++ net/tipc/tipc_cfgsrv.h
@@ -2,7 +2,7 @@
  * net/tipc/tipc_cfgsrv.h: Include file for TIPC configuration service code
  *
  * Copyright (c) 2003-2006, Ericsson AB
- * Copyright (c) 2005, Wind River Systems
+ * Copyright (c) 2005, 2010, Wind River Systems
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -44,7 +44,7 @@
 
 struct sk_buff *tipc_cfg_reply_alloc(int payload_size);
 int tipc_cfg_append_tlv(struct sk_buff *buf, int tlv_type,
-			void *tlv_data, int tlv_data_size);
+			void *tlv_data, size_t tlv_data_size);
 struct sk_buff *tipc_cfg_reply_unsigned_type(u16 tlv_type, u32 value);
 struct sk_buff *tipc_cfg_reply_string_type(u16 tlv_type, char *string);
 
@@ -69,8 +69,8 @@ static inline struct sk_buff *tipc_cfg_reply_ultra_string(char *string)
 }
 
 struct sk_buff *tipc_cfg_do_cmd(u32 orig_node, u16 cmd,
-				const void *req_tlv_area, int req_tlv_space,
-				int headroom);
+				const void *req_tlv_area, u32 req_tlv_space,
+				u32 headroom);
 
 int  tipc_cfg_init(void);
 void tipc_cfg_stop(void);
